Random Circles, Memory Leaks
Magic circles are now once again randomly generated,
provided seed remains None in the config.
Seed can also be set to repeatedly generate the same
circle for draw debugging.

Renderer changed back to default from P2D, unsure
if this was a good choice or not yet, P2D doesn't get
the jittery circles that the default renderer does, but
runs like molasses in wintertime. Either optimization needs
to be done, or the jittering needs to be fixed.

Supposedly some sources claim it may be possible to fix with
drawing each ellipse to an individual pgraphics buffer, but
the new double_ring not only doesn't fix jitter in those
(maybe because three ellipses per), but it leaks memory like mad.
If it can be fixed to not cause jittering there may be some way to
deal with the memory leak issue.